Changed numVisits from int to long Changed the constant temperature to 20

+++ src/main/java/notabot/gametree/GameTreeFactoring.java
@@ -98,7 +98,7 @@ public class GameTreeFactoring extends GameTree {
 	}
 
 
-	public int getNumVisits(){
+	public long getNumVisits(){
 		return root.numVisits;
 	}
